Splits up the components into separate files
